+++ imports/emscripten/emscripten.go
@@ -14,6 +14,7 @@ package emscripten
 
 import (
 	"context"
+	"strings"
 
 	"github.com/tetratelabs/wazero"
 	"github.com/tetratelabs/wazero/api"
@@ -71,15 +72,52 @@ type functionExporter struct{}
 func (functionExporter) ExportFunctions(builder wazero.HostModuleBuilder) {
 	exporter := builder.(wasm.HostFuncExporter)
 	exporter.ExportHostFunc(internal.NotifyMemoryGrowth)
+}
 
-	exporter.ExportHostFunc(internal.NewInvokeFunc("invoke_i", []api.ValueType{i32}, []api.ValueType{i32}))
-	exporter.ExportHostFunc(internal.NewInvokeFunc("invoke_ii", []api.ValueType{i32, i32}, []api.ValueType{i32}))
-	exporter.ExportHostFunc(internal.NewInvokeFunc("invoke_iii", []api.ValueType{i32, i32, i32}, []api.ValueType{i32}))
-	exporter.ExportHostFunc(internal.NewInvokeFunc("invoke_iiii", []api.ValueType{i32, i32, i32, i32}, []api.ValueType{i32}))
-	exporter.ExportHostFunc(internal.NewInvokeFunc("invoke_iiiii", []api.ValueType{i32, i32, i32, i32, i32}, []api.ValueType{i32}))
-	exporter.ExportHostFunc(internal.NewInvokeFunc("invoke_v", []api.ValueType{i32}, nil))
-	exporter.ExportHostFunc(internal.NewInvokeFunc("invoke_vi", []api.ValueType{i32, i32}, nil))
-	exporter.ExportHostFunc(internal.NewInvokeFunc("invoke_vii", []api.ValueType{i32, i32, i32}, nil))
-	exporter.ExportHostFunc(internal.NewInvokeFunc("invoke_viii", []api.ValueType{i32, i32, i32, i32}, nil))
-	exporter.ExportHostFunc(internal.NewInvokeFunc("invoke_viiii", []api.ValueType{i32, i32, i32, i32, i32}, nil))
+type emscriptenFns []*wasm.HostFunc
+
+// InstantiateForModule instantiates a module named "env" populated with any
+// known functions used in emscripten.
+func InstantiateForModule(ctx context.Context, r wazero.Runtime, guest wazero.CompiledModule) (api.Closer, error) {
+	// Create the exporter for the supplied wasm
+	exporter, err := NewFunctionExporterForModule(guest)
+	if err != nil {
+		return nil, err
+	}
+
+	// Instantiate it!
+	env := r.NewHostModuleBuilder("env")
+	exporter.ExportFunctions(env)
+	return env.Instantiate(ctx)
+}
+
+// NewFunctionExporterForModule returns a guest-specific FunctionExporter,
+// populated with any known functions used in emscripten.
+func NewFunctionExporterForModule(guest wazero.CompiledModule) (FunctionExporter, error) {
+	ret := emscriptenFns{}
+	for _, fn := range guest.ImportedFunctions() {
+		importModule, importName, isImport := fn.Import()
+		if !isImport || importModule != "env" {
+			continue // not emscripten
+		}
+		if importName == internal.FunctionNotifyMemoryGrowth {
+			ret = append(ret, internal.NotifyMemoryGrowth)
+			continue
+		}
+		if !strings.HasPrefix(importName, internal.InvokePrefix) {
+			continue // not invoke, and maybe not emscripten
+		}
+
+		hf := internal.NewInvokeFunc(importName, fn.ParamTypes(), fn.ResultTypes())
+		ret = append(ret, hf)
+	}
+	return ret, nil
+}
+
+// ExportFunctions implements FunctionExporter.ExportFunctions
+func (i emscriptenFns) ExportFunctions(builder wazero.HostModuleBuilder) {
+	exporter := builder.(wasm.HostFuncExporter)
+	for _, fn := range i {
+		exporter.ExportHostFunc(fn)
+	}
 }
